Move binder into its own file.
